model をまとめる
